Move actions into their own crate
