+++ crates/modalkit/src/editing/buffer/edit.rs
@@ -486,21 +486,21 @@ mod tests {
 
         // 3r!
         let mov = MoveType::Column(MoveDir1D::Next, false);
-        vctx.replace_char = Some('!'.into());
+        vctx.action.replace = Some('!'.into());
         edit!(ebuf, EditAction::Replace(false), mv!(mov, 3), ctx!(curid, vwctx, vctx), store);
         assert_eq!(ebuf.get_text(), "!!!lo world\na b c d e\nfoo bar baz\n");
         assert_eq!(ebuf.get_leader(curid), Cursor::new(0, 2));
 
         // replace three words ("!", "lo", "world") w/ "Q"
         let mov = MoveType::WordBegin(WordStyle::Little, MoveDir1D::Next);
-        vctx.replace_char = Some('Q'.into());
+        vctx.action.replace = Some('Q'.into());
         edit!(ebuf, EditAction::Replace(false), mv!(mov, 3), ctx!(curid, vwctx, vctx), store);
         assert_eq!(ebuf.get_text(), "!!QQQQQQQQQ\na b c d e\nfoo bar baz\n");
         assert_eq!(ebuf.get_leader(curid), Cursor::new(0, 10));
 
         // replace two lines w/ ":", leaving newlines intact.
         let mov = RangeType::Line;
-        vctx.replace_char = Some(':'.into());
+        vctx.action.replace = Some(':'.into());
         edit!(ebuf, EditAction::Replace(false), range!(mov, 2), ctx!(curid, vwctx, vctx), store);
         assert_eq!(ebuf.get_text(), ":::::::::::\n:::::::::\nfoo bar baz\n");
         assert_eq!(ebuf.get_leader(curid), Cursor::new(1, 8));
@@ -526,8 +526,8 @@ mod tests {
         assert_eq!(get_reg!(store, Register::Unnamed), cell!(CharWise, "world"));
 
         // Test using the named 'a' register ("a).
-        vctx.count = Some(3);
-        vctx.register = Some(Register::Named('a'));
+        vctx.action.count = Some(3);
+        vctx.action.register = Some(Register::Named('a'));
         edit!(ebuf, EditAction::Yank, mv!(mov), ctx!(curid, vwctx, vctx), store);
         assert_eq!(ebuf.get_leader(curid), Cursor::new(0, 6));
 
@@ -537,9 +537,9 @@ mod tests {
         assert_eq!(get_named_reg!(store, 'a'), cell!(CharWise, "world\na b "));
 
         // Append a line to the 'a' register ("A).
-        vctx.count = None;
-        vctx.register = Some(Register::Named('a'));
-        vctx.register_append = true;
+        vctx.action.count = None;
+        vctx.action.register = Some(Register::Named('a'));
+        vctx.action.register_append = true;
         edit!(ebuf, EditAction::Yank, range!(RangeType::Line), ctx!(curid, vwctx, vctx), store);
         assert_eq!(ebuf.get_leader(curid), Cursor::new(0, 6));
 
@@ -552,9 +552,9 @@ mod tests {
         assert_eq!(get_named_reg!(store, 'a'), cell!(LineWise, "world\na b \nhello world\n"));
 
         // The blackhole register ("_) discards the yanked text.
-        vctx.count = None;
-        vctx.register = Some(Register::Blackhole);
-        vctx.register_append = false;
+        vctx.action.count = None;
+        vctx.action.register = Some(Register::Blackhole);
+        vctx.action.register_append = false;
         edit!(ebuf, EditAction::Yank, mv!(mov), ctx!(curid, vwctx, vctx), store);
         assert_eq!(ebuf.get_leader(curid), Cursor::new(0, 6));
 
@@ -680,7 +680,7 @@ mod tests {
         ebuf.set_leader(curid, Cursor::new(0, 7));
 
         // Do a blockwise delete from here to the first word of the third line.
-        vctx.target_shape = Some(TargetShape::BlockWise);
+        vctx.persist.shape = Some(TargetShape::BlockWise);
         edit!(
             ebuf,
             EditAction::Delete,
@@ -717,8 +717,7 @@ mod tests {
         let (mut ebuf, curid, vwctx, mut vctx, mut store) =
             mkfivestr("hello world\na b c d e f\n\n\n1 2 3 4 5 6\n");
 
-        vctx.insert_style = Some(InsertStyle::Insert);
-        vctx.last_column = true;
+        vctx.persist.insert = Some(InsertStyle::Insert);
 
         // Start out at (0, 3).
         ebuf.set_leader(curid, Cursor::new(0, 3));
@@ -758,7 +757,7 @@ mod tests {
 
         // Delete two previous newline characters ("<BS>").
         let mov = MoveType::Column(MoveDir1D::Previous, true);
-        vctx.count = Some(2);
+        vctx.action.count = Some(2);
         edit!(ebuf, EditAction::Delete, mv!(mov), ctx!(curid, vwctx, vctx), store);
         assert_eq!(ebuf.get_leader(curid), Cursor::new(0, 11));
         assert_eq!(ebuf.get_text(), "a b c d e f1 2 3 4 5 6\n");
@@ -840,7 +839,7 @@ mod tests {
         let operation = EditAction::ChangeCase(Case::Toggle);
         let mov = MoveType::Column(MoveDir1D::Next, false);
 
-        vctx.cursor_end = CursorEnd::End;
+        vctx.action.cursor_end = Some(CursorEnd::End);
 
         // Start out at (0, 11), at the start of "eXaMpLE".
         ebuf.set_leader(curid, Cursor::new(0, 11));
@@ -857,7 +856,7 @@ mod tests {
             mkfivestr("hello\nworld\na b c d e\n    word\n");
 
         let op = EditAction::Yank;
-        vctx.target_shape = Some(TargetShape::CharWise);
+        vctx.persist.shape = Some(TargetShape::CharWise);
 
         // Move to (0, 2) to begin.
         ebuf.set_leader(curid, Cursor::new(0, 2));
@@ -896,7 +895,7 @@ mod tests {
             mkfivestr("hello\nworld\na b c d e\n1 2 3 4 5 6");
 
         let op = EditAction::Yank;
-        vctx.target_shape = Some(TargetShape::LineWise);
+        vctx.persist.shape = Some(TargetShape::LineWise);
 
         // Move to (0, 2) to begin.
         ebuf.set_leader(curid, Cursor::new(0, 2));
@@ -931,7 +930,7 @@ mod tests {
         let (mut ebuf, curid, vwctx, mut vctx, mut store) =
             mkfivestr("hello\nworld\na b c d e\n1 2 3 4 5 6");
 
-        vctx.target_shape = Some(TargetShape::BlockWise);
+        vctx.persist.shape = Some(TargetShape::BlockWise);
 
         let mov = MoveType::Line(MoveDir1D::Next);
 
@@ -1113,7 +1112,7 @@ mod tests {
         let opdec = EditAction::ChangeNumber(NumberChange::Decrease(Count::Contextual), false);
         let opinc = EditAction::ChangeNumber(NumberChange::Increase(Count::Contextual), false);
 
-        vctx.count = Some(3);
+        vctx.action.count = Some(3);
         edit!(
             ebuf,
             opinc,
@@ -1124,7 +1123,7 @@ mod tests {
         assert_eq!(ebuf.get_text(), "a 4 b 2 c\nd 3 e 4 f\n");
         assert_eq!(ebuf.get_leader(curid), Cursor::new(0, 2));
 
-        vctx.count = Some(9);
+        vctx.action.count = Some(9);
         edit!(
             ebuf,
             opdec,
@@ -1135,7 +1134,7 @@ mod tests {
         assert_eq!(ebuf.get_text(), "a -5 b 2 c\nd 3 e 4 f\n");
         assert_eq!(ebuf.get_leader(curid), Cursor::new(0, 3));
 
-        vctx.count = Some(1);
+        vctx.action.count = Some(1);
         edit!(
             ebuf,
             opdec,
@@ -1146,7 +1145,7 @@ mod tests {
         assert_eq!(ebuf.get_text(), "a -6 b 2 c\nd 3 e 4 f\n");
         assert_eq!(ebuf.get_leader(curid), Cursor::new(0, 3));
 
-        vctx.count = Some(7);
+        vctx.action.count = Some(7);
         edit!(
             ebuf,
             opinc,
@@ -1159,7 +1158,7 @@ mod tests {
 
         ebuf.set_leader(curid, Cursor::new(1, 4));
 
-        vctx.count = Some(2);
+        vctx.action.count = Some(2);
         edit!(
             ebuf,
             opinc,
@@ -1170,7 +1169,7 @@ mod tests {
         assert_eq!(ebuf.get_text(), "a 1 b 2 c\nd 3 e 6 f\n");
         assert_eq!(ebuf.get_leader(curid), Cursor::new(1, 6));
 
-        vctx.count = Some(3);
+        vctx.action.count = Some(3);
         edit!(
             ebuf,
             opdec,
@@ -1192,12 +1191,12 @@ mod tests {
         let opdec = EditAction::ChangeNumber(NumberChange::Decrease(Count::Contextual), true);
         let opinc = EditAction::ChangeNumber(NumberChange::Increase(Count::Contextual), true);
 
-        vctx.count = Some(5);
+        vctx.action.count = Some(5);
         edit!(ebuf, opinc, &range!(RangeType::Buffer), ctx!(curid, vwctx, vctx), &mut store);
         assert_eq!(ebuf.get_text(), "a 6 b 2 c\nd 13 e 4 f\ng 20 h 6 i\nj 27 k 8 l\n");
         assert_eq!(ebuf.get_leader(curid), Cursor::new(3, 3));
 
-        vctx.count = Some(2);
+        vctx.action.count = Some(2);
         edit!(ebuf, opdec, &range!(RangeType::Buffer), ctx!(curid, vwctx, vctx), &mut store);
         assert_eq!(ebuf.get_text(), "a 4 b 2 c\nd 9 e 4 f\ng 14 h 6 i\nj 19 k 8 l\n");
         assert_eq!(ebuf.get_leader(curid), Cursor::new(3, 3));
